Move the app change receiver to broadcast_receivers
