+++ src/android/app/src/main/kotlin/com/nivisi/leafy_launcher/MainActivity.kt
@@ -18,6 +18,7 @@ import android.util.Base64
 import android.view.View
 import android.view.WindowInsets
 import android.view.WindowManager
+import com.nivisi.leafy_launcher.broadcast_receivers.AppChangeReceiver
 import io.flutter.Log
 import io.flutter.embedding.engine.FlutterEngine
 import io.flutter.plugin.common.EventChannel
