+++ src/android/app/src/main/kotlin/com/nivisi/leafy_launcher/broadcast_receivers/AppChangeReceiver.kt
@@ -1,10 +1,11 @@
-package com.nivisi.leafy_launcher
+package com.nivisi.leafy_launcher.broadcast_receivers
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
+import com.nivisi.leafy_launcher.MainActivity
 
-public class AppChangeReceiver : BroadcastReceiver() {
+class AppChangeReceiver : BroadcastReceiver() {
 
     override fun onReceive(context: Context?, intent: Intent?) {
         val packageName = intent?.data?.encodedSchemeSpecificPart ?: return
